requirements.txt
Flask==3.0.3
google-cloud-aiplatform==1.84.0
google-cloud-speech ==2.27.0
google-cloud-language==2.13.1
gunicorn==22.0.0
Werkzeug==3.1.3